refractor injiweb changes to v1 to have common version for mobile and web Signed-off-by: Challa Rao <47347156+challabeehyv@users.noreply.github.com>
