+++ pom.xml
@@ -89,8 +89,8 @@
         <bouncycastle.version>1.70</bouncycastle.version>
         <jwt.version> 3.8.1</jwt.version>
         <google.zxing.version>3.3.3</google.zxing.version>
-        <itextcore.version>7.2.0</itextcore.version>
-        <itexthtml2pdf.version>2.0.0</itexthtml2pdf.version>
+        <itextcore.version>8.0.3</itextcore.version>
+        <itexthtml2pdf.version>5.0.3</itexthtml2pdf.version>
         <itext.version>5.5.13</itext.version>
         <commons-math3>3.6.1</commons-math3>
         <commons-lang3>3.7</commons-lang3>
@@ -98,8 +98,6 @@
         <commons-codec>1.15</commons-codec>
         <swagger.version>2.9.2</swagger.version>
         <springdoc.version>1.5.10</springdoc.version>
-        <flying-saucer-pdf-version>9.3.1</flying-saucer-pdf-version>
-        <josup-version>1.16.2</josup-version>
         <jjwt-version>0.9.1</jjwt-version>
         <sonar.coverage.exclusions>
             **/constant/**,**/config/**,**/httpfilter/**,**/cache/**,**/dto/**,**/entity/**,**/model/**,**/exception/**,**/repository/**,**/security/**,**/*Config.java,**/*BootApplication.java,**/*VertxApplication.java,**/cbeffutil/**,**/core.http/**,**/util/**
@@ -147,20 +145,6 @@
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-webflux</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-thymeleaf</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.xhtmlrenderer</groupId>
-            <artifactId>flying-saucer-pdf-openpdf</artifactId>
-            <version>${flying-saucer-pdf-version}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.jsoup</groupId>
-            <artifactId>jsoup</artifactId>
-            <version>${josup-version}</version>
         </dependency>
         <dependency>
             <groupId>org.springframework.boot</groupId>
