+++ src/main/java/io/mosip/mimoto/dto/mimoto/CredentialIssuerWellKnownResponse.java
@@ -1,4 +1,4 @@
-package io.mosip.mimoto.dto.v2;
+package io.mosip.mimoto.dto.mimoto;
 
 import lombok.Data;
