+++ src/main/resources/application-local.properties
@@ -1,7 +1,7 @@
 
 # MOSIP
 mosipbox.public.url=https://api-internal.dev2.mosip.net
-public.internet.url=http://localhost:8088
+public.internet.url=https://api.dev2.mosip.net
 public.url=https://api-internal.dev2.mosip.net
 keycloak.internal.url=https://iam.dev2.mosip.net/
 
@@ -160,8 +160,7 @@ CREATEDATASHARE=${mosipbox.public.url}/v1/datashare/create
 DECRYPTPINBASSED=${mosipbox.public.url}/v1/keymanager/decryptWithPin
 
 # config.server.file.storage.uri=${spring.cloud.config.uri}/${spring.application.name}/${spring.profiles.active}/${spring.cloud.config.label}/
-#config.server.file.storage.uri=${mosipbox.public.url}/config/print/mz/develop2-v3/
-config.server.file.storage.uri=https://challabeehyv.github.io/mimoto-config/
+config.server.file.storage.uri=${mosipbox.public.url}/config/print/mz/develop2-v3/
 
 # websub authentication - auth adapter
 mosip.iam.adapter.clientid=mpartner-default-mobile
@@ -200,44 +199,13 @@ wallet.binding.partner.api.key=Aci9jg28B8mO_LDfDXo3ZTp5_HKgEMun2tYyHCa1e8k
 #wallet.binding.partner.api.key=14378
 
 # OpenID
-mosip.openid.issuers=v2-issuers-config.json
-mosip.openid.v2.issuer.credentialSupported=/wellKnownIssuer/Insurance.json
-mosip.openid.v2.htmlTemplate=CredentialTemplate.html
+mosip.openid.issuers=mimoto-issuers-config.json
+mosip.openid.issuer.credentialSupported=/wellKnownIssuer/Insurance.json
+mosip.openid.htmlTemplate=CredentialTemplate.html
 mosip.oidc.client.assertion.type=urn:ietf:params:oauth:client-assertion-type:jwt-bearer
 mosip.oidc.p12.filename=oidckeystore.p12
 mosip.oidc.p12.password=5YBx6QT2wbY8Ls6w
 mosip.oidc.p12.path=certs/
-mosip.oidc.p12.alias=mpartner-default-mimotooidc
-
-mosip.esignet.host=http://localhost:8088
-
-# Inji properties
-mosip.inji.allowedAuthType=demo,otp,bio-Finger,bio-Iris,bio-Face
-mosip.inji.allowedEkycAuthType=demo,otp,bio-Finger,bio-Iris,bio-Face
-mosip.inji.allowedInternalAuthType=otp,bio-Finger,bio-Iris,bio-Face
-mosip.inji.faceSdkModelUrl=null
-# model download maximum retry
-mosip.inji.modelDownloadMaxRetry=10
-# maximum number of retry for downloading vc
-mosip.inji.vcDownloadMaxRetry=10
-# pool interval in milli seconds
-mosip.inji.vcDownloadPoolInterval=6000
-# validate binding audience url to be sent in token
-mosip.inji.audience=ida-binding
-# issuer to be sent in token
-mosip.inji.issuer=residentapp
-# warning screen domain name -- mimoto host
-mosip.inji.warningDomainName=http://localhost:8088/
-# inji documentation url
-mosip.inji.aboutInjiUrl=https://docs.mosip.io/inji
-# minimum storage space required for making audit entry in MB
-mosip.inji.minStorageRequiredForAuditEntry=2
-# minimum storage space required for downloading / receiving vc in MB
-mosip.inji.minStorageRequired=5
-#timeout for vc download api via openid4vci flow in milliseconds
-mosip.inji.openId4VCIDownloadVCTimeout=30000
-
-mosip.security.origins=*
