+++ src/main/resources/templates/CredentialTemplate.html
@@ -7,34 +7,38 @@
 </head>
 
 <body #if ($keyFontColor) style="color: $keyFontColor;" #end>
-<div #if ($bgColor) style="background: $bgColor; border: 2px black solid; border-radius: 15px; padding: 10px;" #else style="border: 2px black solid; border-radius: 15px; padding: 10px;" #end>
+<div #if ($bgColor) style="background: $bgColor; border: 1px $keyFontColor solid; border-radius: 15px; padding: 10px;" #else style="border: 2px black solid; border-radius: 15px; padding: 10px;" #end>
     <div style="text-align: center; padding-top: 10px; font-weight: bold;">
-        <img style="width: 20px; margin-right: 10px; vertical-align: middle;" src="$titleLogo" alt="Logo" />
         <span style="font-size: 16px; vertical-align: middle;">$titleName</span>
     </div>
     <div style="margin: 20px;">
-        <img style="width: 100px; float: left; margin-right: 100px;" src="https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcRe3oPvKsA05otgZYGFZmxk5WHLYTFKWOFaNA" alt="Logo" />
-        <div style="overflow: auto;">
+        <div style="float: left; margin-right: 50px; height:100px;width:100px;background-color:white;border-radius:10px;border:1px solid grey">
+            <img style="margin: 25px;width: 50px; height:50px;" src="https://raw.githubusercontent.com/challabeehyv/mimoto-config/main/user-icon.png" alt="Logo" />
+        </div>
+        <div style="overflow: auto; float: left">
             #foreach($entry in $headerProperties.entrySet())
             <div style="margin-bottom: 10px;">
                 <div style="font-size: 16px; margin-bottom: 5px;">$entry.key</div>
-                <div style="font-weight: bold; font-size: 16px; margin-left: 10px;">$entry.value</div>
+                <div style="font-weight: bold; font-size: 16px;">$entry.value</div>
             </div>
             #end
         </div>
+        <div style="overflow:auto;text-align:right">
+            <img style="width:75px;height:75px;" src="https://raw.githubusercontent.com/challabeehyv/mimoto-config/main/QR_Code.png"/>
+        </div>
     </div>
-    <div style="margin-left: 20px;">
+    <div style="margin-left: 20px; padding-top: 20px">
         #foreach($entry in $rowProperties.entrySet())
         #if($entry.value)
-        <div style="margin-bottom: 10px; float: left; width: 250px;">
+        <div style="margin-bottom: 10px; float: left; width: 300px;">
             <div style="font-size: 16px; margin-bottom: 5px;">$entry.key</div>
             <div style="font-weight: bold; font-size: 16px;">$entry.value</div>
         </div>
         #end
         #end
     </div>
-    <div style="text-align: right; #if($rowPropertiesMargin) margin-top: ${rowPropertiesMargin}px; #end">
-        <img style="width: 100px;" src="$logoUrl" alt="Logo" />
+    <div style="text-align: right;margin-left: 50px; margin-right: 20px; #if($rowPropertiesMargin) margin-top: ${rowPropertiesMargin}px; #end">
+        <img style="width: 75px;" src="$logoUrl" alt="Logo" />
     </div>
 </div>
 </body>
